pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "LangChain agent travel planner using Google Maps APIs"
requires-python = ">=3.10"
dependencies = [
  "langchain>=1.2.5",
  "langchain-openai>=0.3.0",
  "langchain-core>=0.3.0",
  "langgraph>=1.0.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "requests>=2.31.0",
]

[project.scripts]
travel-agent = "agent_app_fixed:main"

[tool.uv]
# uv reads dependencies from [project] above.
